@@ -6,18 +6,18 @@ version: 1.9.0
 
 environment:
   sdk: '>=3.0.0 <4.0.0'
-  flutter: '3.19.3'
+  flutter: '3.22.2'
 
 dependencies:
-  cloud_firestore: ^4.0.0
+  cloud_firestore: ^5.1.0
   firebase_core_platform_interface: ^5.0.0
-  firebase_crashlytics: ^3.0.0
-  firebase_core: ^2.0.0
-  firebase_auth: ^4.0.0
-  firebase_dynamic_links: ^5.0.0
-  firebase_app_check: ^0.2.0
+  firebase_crashlytics: ^4.0.3
+  firebase_core: ^3.2.0
+  firebase_auth: ^5.1.2
+  firebase_dynamic_links: ^6.0.3
+  firebase_app_check: ^0.3.0+3
   google_sign_in: ^6.0.0
-  google_mobile_ads: ^5.0.0
+  google_mobile_ads: ^5.1.0
   cupertino_icons: ^1.0.2
   enum_to_string: ^2.0.1
   flutter_markdown: ^0.7.0
@@ -28,13 +28,13 @@ dependencies:
   font_awesome_flutter: ^10.0.0
   google_fonts: ^6.0.0
   http: ^1.1.2
-  intl: ^0.18.1
+  intl: ^0.19.0
   in_app_purchase: ^3.0.7
   package_info_plus : ^8.0.0
   provider: ^6.0.0
   url_launcher: ^6.0.2
-  syncfusion_flutter_charts: ^24.1.43+1
-  syncfusion_flutter_gauges: ^24.1.43
+  syncfusion_flutter_charts: ^26.1.42
+  syncfusion_flutter_gauges: ^26.1.42
   infinite_scroll_pagination: ^4.0.0
   flutter:
     sdk: flutter
@@ -55,7 +55,8 @@ dev_dependencies:
   mockito: ^5.0.0
   network_image_mock: ^2.0.1
   build_runner: ^2.1.2
-  flutter_lints: ^3.0.1
+  flutter_lints: ^4.0.0
+  flutter_gen_runner: ^5.6.0
 
 
 flutter:
