@@ -104,10 +104,10 @@ class PlayerColorExplanationDialog extends StatelessWidget {
         ElevatedButton.icon(
           key: const ValueKey('closeButton'),
           style: ButtonStyle(
-              backgroundColor: MaterialStateProperty.all<Color>(Colors.white),
-              foregroundColor: MaterialStateProperty.all<Color>(
+              backgroundColor: WidgetStateProperty.all<Color>(Colors.white),
+              foregroundColor: WidgetStateProperty.all<Color>(
                   Theme.of(context).primaryColor),
-              shape: MaterialStateProperty.all<OutlinedBorder>(
+              shape: WidgetStateProperty.all<OutlinedBorder>(
                   RoundedRectangleBorder(
                       borderRadius: BorderRadius.circular(
                           CustomProperties.borderRadius)))),
