@@ -37,11 +37,11 @@ class SpecialRoundDialog extends StatelessWidget {
         children: [
           ElevatedButton(
             style: ButtonStyle(
-              backgroundColor: MaterialStateProperty.all<Color>(
+              backgroundColor: WidgetStateProperty.all<Color>(
                   Theme.of(context).primaryColor),
               foregroundColor:
-                  MaterialStateProperty.all<Color>(Theme.of(context).cardColor),
-              shape: MaterialStateProperty.all<OutlinedBorder>(
+                  WidgetStateProperty.all<Color>(Theme.of(context).cardColor),
+              shape: WidgetStateProperty.all<OutlinedBorder>(
                 RoundedRectangleBorder(
                   borderRadius:
                       BorderRadius.circular(CustomProperties.borderRadius),
@@ -58,11 +58,11 @@ class SpecialRoundDialog extends StatelessWidget {
           ),
           ElevatedButton(
             style: ButtonStyle(
-              backgroundColor: MaterialStateProperty.all<Color>(
+              backgroundColor: WidgetStateProperty.all<Color>(
                   Theme.of(context).primaryColor),
               foregroundColor:
-                  MaterialStateProperty.all<Color>(Theme.of(context).cardColor),
-              shape: MaterialStateProperty.all<OutlinedBorder>(
+                  WidgetStateProperty.all<Color>(Theme.of(context).cardColor),
+              shape: WidgetStateProperty.all<OutlinedBorder>(
                 RoundedRectangleBorder(
                   borderRadius:
                       BorderRadius.circular(CustomProperties.borderRadius),
@@ -79,11 +79,11 @@ class SpecialRoundDialog extends StatelessWidget {
           ),
           ElevatedButton(
             style: ButtonStyle(
-              backgroundColor: MaterialStateProperty.all<Color>(
+              backgroundColor: WidgetStateProperty.all<Color>(
                   Theme.of(context).colorScheme.error),
               foregroundColor:
-                  MaterialStateProperty.all<Color>(Theme.of(context).cardColor),
-              shape: MaterialStateProperty.all<OutlinedBorder>(
+                  WidgetStateProperty.all<Color>(Theme.of(context).cardColor),
+              shape: WidgetStateProperty.all<OutlinedBorder>(
                 RoundedRectangleBorder(
                   borderRadius:
                       BorderRadius.circular(CustomProperties.borderRadius),
