@@ -124,11 +124,11 @@ class _UserScreenState extends State<UserScreen>
                         AppLocalizations.of(context)!.youDontHaveAnyPlayer),
                 ElevatedButton.icon(
                   style: ButtonStyle(
-                    backgroundColor: MaterialStateProperty.all<Color>(
+                    backgroundColor: WidgetStateProperty.all<Color>(
                         Theme.of(context).primaryColor),
-                    foregroundColor: MaterialStateProperty.all<Color>(
+                    foregroundColor: WidgetStateProperty.all<Color>(
                         Theme.of(context).cardColor),
-                    shape: MaterialStateProperty.all<OutlinedBorder>(
+                    shape: WidgetStateProperty.all<OutlinedBorder>(
                       RoundedRectangleBorder(
                         borderRadius: BorderRadius.circular(
                           CustomProperties.borderRadius,
@@ -444,10 +444,10 @@ class _AppBarTitle extends StatelessWidget {
         FilledButton(
           style: ButtonStyle(
             backgroundColor:
-                MaterialStateProperty.all<Color>(Theme.of(context).cardColor),
-            foregroundColor: MaterialStateProperty.all<Color>(
-                Theme.of(context).primaryColor),
-            shape: MaterialStateProperty.all<OutlinedBorder>(
+                WidgetStateProperty.all<Color>(Theme.of(context).cardColor),
+            foregroundColor:
+                WidgetStateProperty.all<Color>(Theme.of(context).primaryColor),
+            shape: WidgetStateProperty.all<OutlinedBorder>(
               RoundedRectangleBorder(
                 borderRadius: BorderRadius.circular(
                   CustomProperties.borderRadius,
